Исправил замечания и сделал добавки
Расширил DTO модели. Создал Мапперы моделей и DTO

+++ HealthDiary/MetricService.BLL/Services/UserService.cs
@@ -1,6 +1,7 @@
-﻿using MetricService.BLL.Dto;
+﻿using MetricService.BLL.DTO;
 using MetricService.BLL.Exceptions;
 using MetricService.BLL.Interfaces;
+using MetricService.BLL.Mappers;
 using MetricService.DAL.Interfaces;
 using System.Security.Claims;
 
@@ -15,17 +16,25 @@ namespace MetricService.BLL.Services
         /// <summary>
         /// Удаление профиля пользователя
         /// </summary>
-        /// <param name="userId"></param>
-        /// <returns>true - в случае успеха</returns>
-        /// <exception cref="ViolationAccessException">Возникает при нарушении уровня доступа к чужим данным</exception>        
-        public Task<bool> DeleteProfileAsync(int userId)
+        /// <param name="userId"></param>       
+        /// <exception cref="ViolationAccessException">Возникает при нарушении уровня доступа к чужим данным</exception>    
+        /// <exception cref="IncorrectOrEmptyResultException">Возникает если указанный пользователь не существует</exception>   
+        public async Task DeleteProfileAsync(int userId)
         {
+            if (await _repository.GetByIdAsync(userId) == null)
+                throw new IncorrectOrEmptyResultException("Указанный пользователь не существует", new Dictionary<object, object>()
+                {
+                    { "userId", userId }
+                });
+
             if (!_authorizationService.IsInRole("Admin") && userId != Common.Common.GetAuthorId(_authorizationService))
             {
-                throw new ViolationAccessException("Вам разрешено удалить только свой профиль", Common.Common.GetAuthorId(_authorizationService), userId, _repository.Name);
+                throw new ViolationAccessException("Вам разрешено удалить только свой профиль", 
+                    Common.Common.GetAuthorId(_authorizationService), userId, _repository.Name);
             }
 
-            return _repository.DeleteAsync(userId);
+            await _repository.DeleteAsync(userId);
+
         }
 
         /// <summary>
@@ -42,17 +51,9 @@ namespace MetricService.BLL.Services
                 throw new ViolationAccessException("Только администраторам разрешено просматривать список пользователей", Common.Common.GetAuthorId(_authorizationService), 0, _repository.Name);
             }
 
-            var users = (await _repository.GetAllAsync()).Skip((pageNum - 1) * pageSize).Take(pageSize);
+            var users = (await _repository.GetAllAsync()).Skip((pageNum - 1) * pageSize).Take(pageSize).ToUserDTO();                     
 
-            var usersDTO = new List<UserDTO>();
-            if (users.Any())
-            {
-                foreach (var user in users)
-                {
-                    usersDTO.Add(CreateUserDTO(user)!);
-                }
-            }
-            return usersDTO;
+            return users;
         }
 
         /// <summary>
@@ -60,40 +61,58 @@ namespace MetricService.BLL.Services
         /// </summary>
         /// <param name="userId">идентификатор пользователя</param>
         /// <returns>Модель DTO</returns>
-        /// <exception cref="ViolationAccessException">Возникает при нарушении уровня доступа к чужим данным</exception>        
-        public async Task<UserDTO?> GetUserByIdAsync(int userId)
+        /// <exception cref="ViolationAccessException">Возникает при нарушении уровня доступа к чужим данным</exception>  
+        /// <exception cref="IncorrectOrEmptyResultException">Возникает когда пользователь с заданным ИД не найден</exception> 
+        public async Task<UserDTO> GetUserByIdAsync(int userId)
         {
+            var findUser = await _repository.GetByIdAsync(userId)??
+                throw new IncorrectOrEmptyResultException("Пользователь не найден",
+                new Dictionary<object, object>()
+                {
+                    {"userId", userId},
+                });
+
             if (!_authorizationService.IsInRole("Admin") && userId != Common.Common.GetAuthorId(_authorizationService))
             {
-                throw new ViolationAccessException("Вы не можете просматривать профиль другого пользователя", Common.Common.GetAuthorId(_authorizationService), userId, _repository.Name);
+                throw new ViolationAccessException("Вы не можете просматривать профиль другого пользователя", 
+                    Common.Common.GetAuthorId(_authorizationService), userId, _repository.Name);
             }
 
-            return CreateUserDTO(await _repository.GetByIdAsync(userId));
+            return findUser.ToUserDTO();
         }
 
         /// <summary>
         /// Создание профиля пользователя
         /// </summary>
         /// /// <param name="author"></param>
-        /// <param name="userDTO"></param>
-        /// <returns>true - в случае успеха</returns>
+        /// <param name="userDTO"></param>        
         /// <exception cref="ViolationAccessException">Возникает при нарушении уровня доступа к чужим данным</exception>
         /// <exception cref="ValidateModelException">Возникает когда данные содержат не корректные данные</exception>
-        public async Task<bool> CreateProfileAsync(UserDTO userDTO)
+        ///  <exception cref="IncorrectOrEmptyResultException">Возникает когда пользователь уже зарегистрирован</exception>
+        public async Task CreateProfileAsync(UserDTO userDTO)
         {
+            
+            if (await _repository.GetByIdAsync(userDTO.Id) != null)
+                throw new IncorrectOrEmptyResultException("Пользователь уже зарегистрирован",
+                    new Dictionary<object, object>()
+                    {
+                        {"userDTO", userDTO}
+                    });
+
             if (!_authorizationService.IsInRole("Admin") && userDTO.Id != Common.Common.GetAuthorId(_authorizationService))
             {
-                throw new ViolationAccessException("Вы не можете создать данные другого пользователя", Common.Common.GetAuthorId(_authorizationService), userDTO.Id, _repository.Name);
+                throw new ViolationAccessException("Вы не можете создать данные другого пользователя", 
+                    Common.Common.GetAuthorId(_authorizationService), userDTO.Id, _repository.Name);
             }
 
-            Domain.Models.User? user = CreateUser(userDTO);
+            Domain.Models.User? user = userDTO.ToUser();
 
-            if (!_validator.Validate(user!, out IDictionary<string, string> errorList))
+            if (!_validator.Validate(user!, out Dictionary<string, string> errorList))
             {
                 throw new ValidateModelException("Некорректные данные пользователя", errorList);
             }
 
-            return await _repository.CreateAsync(user);
+            await _repository.CreateAsync(user);
         }
 
         /// <summary>
@@ -104,14 +123,19 @@ namespace MetricService.BLL.Services
         /// <returns>true - в случае успеха</returns>
         /// <exception cref="ViolationAccessException">Возникает при нарушении уровня доступа к чужим данным</exception>
         /// <exception cref="ValidateModelException">Возникает когда данные содержат не корректные данные</exception>
-        public async Task<bool> UpdateProfileAsync(UserDTO userDTO)
+        public async Task UpdateProfileAsync(UserDTO userDTO)
         {
-            var findUser = await _repository.GetByIdAsync(userDTO.Id);
-            if (findUser == null) return false;
+            var findUser = await _repository.GetByIdAsync(userDTO.Id) ?? 
+                throw new IncorrectOrEmptyResultException("Пользователь не зарегистрирован",
+                new Dictionary<object, object>()
+                {
+                    {"userDTO", userDTO },
+                });           
 
             if (!_authorizationService.IsInRole("Admin") && findUser.Id != Common.Common.GetAuthorId(_authorizationService))
             {
-                throw new ViolationAccessException("Вы не можете изменять данные другого пользователя", Common.Common.GetAuthorId(_authorizationService), findUser.Id, _repository.Name);
+                throw new ViolationAccessException("Вы не можете изменять данные другого пользователя", 
+                    Common.Common.GetAuthorId(_authorizationService), findUser.Id, _repository.Name);
             }
 
             findUser.Id = userDTO.Id;
@@ -119,49 +143,13 @@ namespace MetricService.BLL.Services
             findUser.Height = userDTO.Height;
             findUser.DateOfBirth = userDTO.DateOfBirth;
 
-            if (!_validator.Validate(findUser!, out IDictionary<string, string> errorList))
+            if (!_validator.Validate(findUser!, out Dictionary<string, string> errorList))
             {
                 throw new ValidateModelException("Некорректные данные пользователя", errorList);
 
             }
-            
-            return await _repository.UpdateAsync(findUser);
+
+            await _repository.UpdateAsync(findUser);
         }
-
-        /// <summary>
-        /// Создание UserDTO по модели пользователя
-        /// </summary>
-        /// <param name="user"></param>
-        /// <returns>Модель DTO</returns>
-        private static UserDTO? CreateUserDTO(Domain.Models.User? user)
-        {
-            if (user == null) return null;
-            return new UserDTO
-            {
-                Id = user.Id,
-                DateOfBirth = user.DateOfBirth,
-                Height = user.Height,
-                Weight = user.Weight,
-                Age = user.Age,
-            };
-        }
-
-
-        /// <summary>
-        /// Создание модели из DTO
-        /// </summary>
-        /// <param name="userDTO"></param>
-        /// <returns>Модлель</returns>
-        private static Domain.Models.User CreateUser(UserDTO userDTO)
-        {
-            return new Domain.Models.User
-            {
-                Id = userDTO.Id,
-                DateOfBirth = userDTO.DateOfBirth,
-                Height = userDTO.Height,
-                Weight = userDTO.Weight
-            };
-        }
-
     }
 }
